Add `<InlineMessage>`
This component is similar to `<Message>`, however you provide a translation, in any
locale, as a child of the component. This could be more useful than `<Message>` since,
when developing a page, you can see what the actual content looks like, rather than just
the `pointer` to where the content lives.

We'll try this out for a while, and if successful, replace `<Message>` for this.